Working Login page
1) The login page is updated to use html facade as interface.
2) The Routes.php is updated to include the returnPage method from
LoginController controller.
3)  The controller page is updated with Validator, Input and Redirect
illuminator which we have to do manually. We have to run following code
to get the illuminator package so that we can use the form facade and
html facade functionality.
composer require illuminate/html

4) The config/app.php is also the part of the update we have to make
after downloading the illuminate package. We have to update the code to
include the Form, HTML and Input facade.

5) The errors list have been added to separate page called
list.blade.php
on the errors folder
6) Test carried out using the username and password provided in the
seeder and was successful.

7)Validation test performed without adding any contents on the username
and password field which generated error as it should.