Modification of tables
Changed the style of the tables/database, so now the root causes have a
separate table with ids to be called for instead of being entered
manually each time, and tables with root cause being used have been
altered to reflect this. Migrations and seeders have also been tweaked
to match up with the new changes.